Split dump substate comand
