+++ cmd/util-db/dump_substate/dump_substate.go
@@ -1,33 +1,16 @@
-// Copyright 2024 Fantom Foundation
-// This file is part of Aida Testing Infrastructure for Sonic
-//
-// Aida is free software: you can redistribute it and/or modify
-// it under the terms of the GNU Lesser General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-//
-// Aida is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-// GNU Lesser General Public License for more details.
-//
-// You should have received a copy of the GNU Lesser General Public License
-// along with Aida. If not, see <http://www.gnu.org/licenses/>.
-
-package db
+package dump_substate
 
 import (
 	"fmt"
-
 	"github.com/0xsoniclabs/aida/utildb"
 	"github.com/0xsoniclabs/aida/utils"
 	"github.com/0xsoniclabs/substate/db"
 	"github.com/urfave/cli/v2"
 )
 
-// SubstateDumpCommand returns content in substates in json format
-var SubstateDumpCommand = cli.Command{
-	Action:    substateDumpAction,
+// Command returns content in substates in json format
+var Command = cli.Command{
+	Action:    dumpSubstateAction,
 	Name:      "dump-substate",
 	Usage:     "returns content in substates in json format",
 	ArgsUsage: "<blockNumFirst> <blockNumLast>",
@@ -44,8 +27,8 @@ The aida-vm dump command requires two arguments:
 last block of the inclusive range of blocks to replay transactions.`,
 }
 
-// substateDumpAction prepares config and arguments before SubstateDumpAction
-func substateDumpAction(ctx *cli.Context) error {
+// dumpSubstateAction prepares config and arguments before SubstateDumpAction
+func dumpSubstateAction(ctx *cli.Context) error {
 	var err error
 
 	cfg, err := utils.NewConfig(ctx, utils.BlockRangeArgs)
